requirements.txt
annotated-types==0.6.0
anyio==3.7.1
black==23.11.0
certifi==2023.7.22
click==8.1.7
distro==1.8.0
flake8==6.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
openai==1.2.4
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
psycopg==3.1.12
pycodestyle==2.11.1
pydantic==2.5.0
pydantic_core==2.14.1
pyflakes==3.1.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
